requirements.txt
tifffile
tqdm
pillow
gradio
timm==1.0.3
einops==0.6.0
opencv-python
numpy==1.26.4
bioimageio.core==0.5.9
scipy
scikit-image
pytorch_msssim
warmup_scheduler
tensorboard